requirements.txt
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
redis>=5.0.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
httpx>=0.26.0
python-multipart>=0.0.6
openai>=1.12.0
tenacity>=8.2.3
python-dateutil>=2.8.2
jinja2>=3.1.3
